requirements/common.txt
Flask==0.12.2
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
MySQL-python==1.2.5
Flask-Script==2.0.5
python-dotenv==0.6.4
Flask-RESTful==0.3.6
